Load NPC data from maps, Fix Select() timeout, Moved Refresh packet to a function
git-svn-id: svn://svn.eoserv.net/eoserv/trunk@99 da1c5d5e-72c9-4536-8af9-af1c9134b0e6

+++ src/socket.cpp
@@ -285,7 +285,7 @@ void Client::Tick(double timeout)
 {
 	fd_set read_fds, write_fds, except_fds;
 	long tsecs = long(timeout);
-	timeval timeout_val = {tsecs, (long(timeout) - tsecs)*1000000};
+	timeval timeout_val = {tsecs, long((timeout - double(tsecs))*1000000)};
 
 	FD_ZERO(&read_fds);
 	FD_ZERO(&write_fds);
